backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.27.2
openai==1.58.1
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.21.1
redis==5.2.1
python-socketio==5.12.0
celery==5.4.0
supabase==2.10.0
# RAG dependencies for PDF extraction
pymupdf==1.24.0
pdf2image==1.17.0  # Optional: requires poppler-utils
sentence-transformers==2.2.2
numpy==1.26.4
faiss-cpu==1.7.4
Pillow==10.2.0
